// src/modules/atacado-repository/wholesale.repository.types.ts
import type { Cliente } from "@/@types/atacado/Cliente";
import type { Cobranca } from "@/@types/atacado/Cobranca";
import type { Fatura } from "@/@types/atacado/Fatura";
import type { ItemNFCom } from "@/@types/atacado/ItemNFCom";
import type { NFCom } from "@/@types/atacado/NFCom";
import type { Parceiro } from "@/@types/atacado/Parceiro";
import type { PlanoDeServico } from "@/@types/atacado/PlanoDeServico";

type OmitAutoGenerated<T> = Omit<
	T,
	"id" | "createdAt" | "updatedAt" | "createdById" | "updatedById"
>;

interface ClientsFilter {
	readonly partnerId: string | number;
	readonly activeLineStatus?: boolean;
}

interface AtacadoRepository {
	findParceiroById(id: string | number): Promise<Parceiro>;

	findClientesAtivosByParceiroId(
		filter: ClientsFilter,
	): Promise<readonly Cliente[]>;

	findAllPlanosDeServico(): Promise<readonly PlanoDeServico[]>;

	createFatura(data: OmitAutoGenerated<Fatura>): Promise<Fatura>;

	createCobranca(data: OmitAutoGenerated<Cobranca>): Promise<Cobranca>;

	createNFCom(
		cobrancaId: string | number,
		data: OmitAutoGenerated<NFCom>,
	): Promise<NFCom>;

	createItemNFCom(
		notaFiscalId: string | number,
		data: OmitAutoGenerated<ItemNFCom>,
	): Promise<ItemNFCom>;

	deleteFatura(id: string | number): Promise<void>;

	deleteCobranca(id: string | number): Promise<void>;

	deleteNFCom(id: string | number): Promise<void>;

	deleteItemNFCom(id: string | number): Promise<void>;
}

export type { AtacadoRepository, ClientsFilter, OmitAutoGenerated };